refactor: move logic to separate library
